refacto handle applicants file
